requirements.txt
flask
numpy
google-generativeai
python-dotenv
